Using templates to fast compute % and / for power of 2 values in bGrid.

+++ libNeonDomain/include/Neon/domain/details/bGrid/bField.h
@@ -18,7 +18,7 @@
 namespace Neon::domain::details::bGrid {
 
 
-template <typename T, int C, int8_t memBlockSizeX, int8_t memBlockSizeY, int8_t memBlockSizeZ, int8_t userBlockSizeX, int8_t userBlockSizeY, int8_t userBlockSizeZ>
+template <typename T, int C, uint32_t memBlockSizeX, uint32_t memBlockSizeY, uint32_t memBlockSizeZ, uint32_t userBlockSizeX, uint32_t userBlockSizeY, uint32_t userBlockSizeZ>
 class bField : public Neon::domain::interface::FieldBaseTemplate<T,
                                                                  C,
                                                                  bGrid<memBlockSizeX, memBlockSizeY, memBlockSizeZ, userBlockSizeX, userBlockSizeY, userBlockSizeZ>,
